+++ enforcer/integration_test/container_integration_test.rs
@@ -101,6 +101,7 @@ async fn verify_write(bundle_path: &str) -> anyhow::Result<(NamedTempFile, Vec<u
         network: NetworkOptions::default(),
         env: vec![],
         run_isolate_as_unprivileged: false,
+        seccomp_profile: Default::default(),
     };
     container.start(&opts).await?;
 
@@ -143,6 +144,7 @@ async fn verify_read(bundle_path: &str, file: &NamedTempFile) -> anyhow::Result<
         network: NetworkOptions::default(),
         env: vec![],
         run_isolate_as_unprivileged: false,
+        seccomp_profile: Default::default(),
     };
     container.start(&opts).await?;
 
@@ -184,6 +186,7 @@ async fn verify_read_only(bundle_path: &str, file: &NamedTempFile) -> anyhow::Re
         network: NetworkOptions::default(),
         env: vec![],
         run_isolate_as_unprivileged: false,
+        seccomp_profile: Default::default(),
     };
     container.start(&opts).await?;